requirements.txt
# Required packages for program
pynacl


# Required packages for testing
pytest
spyder-unittest
